requirements.txt
pyobjc-framework-Quartz
bleak